refactor: move date utility
